+++ lang/syn/src/codegen/accounts/mod.rs
@@ -22,7 +22,7 @@ pub fn generate(accs: &AccountsStruct) -> proc_macro2::TokenStream {
     let __client_accounts_mod = __client_accounts::generate(accs);
     let __cpi_client_accounts_mod = __cpi_client_accounts::generate(accs);
 
-    quote! {
+    let ret = quote! {
         #impl_try_accounts
         #impl_to_account_infos
         #impl_to_account_metas
@@ -30,7 +30,21 @@ pub fn generate(accs: &AccountsStruct) -> proc_macro2::TokenStream {
 
         #__client_accounts_mod
         #__cpi_client_accounts_mod
+    };
+
+    #[cfg(feature = "idl-build")]
+    {
+        #![allow(warnings)]
+        let no_docs = crate::idl::build::get_no_docs();
+        let idl_gen_impl = crate::idl::build::gen_idl_gen_impl_for_accounts_strct(&accs, no_docs);
+        return quote! {
+            #ret
+            #idl_gen_impl
+        };
     }
+
+    #[allow(unreachable_code)]
+    ret
 }
 
 fn generics(accs: &AccountsStruct) -> ParsedGenerics {